added query argument to script

+++ src/config.rs
@@ -20,4 +20,4 @@ pub fn get_delta(db_size: u32, delta: Option<f32>) -> f32 {
 pub const PROVER_ADDRESS: &str = "127.0.0.1";
 pub const PROVER_PORT: &str = "10020";
 
-pub type DataT = u32;
+pub type DataT = u16;
